+++ CHANGELOG.md
@@ -5,7 +5,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](#) - ????-??-??
 
-## [2.4.0](v.2.4.0) - 2017-07-04
+## [2.5.0](v2.5.0) - 2017-07-12
+### Added
+- Group BankAccounts display now as well
+- Print-view for all characters and bank accounts
+
+### Changed
+- Totes fixed BankAccounts
+
+### Removed
+- Rake DB tasks in favor of the shell script from 2.2.0 / 460803cb
+
+## [2.4.0](v2.4.0) - 2017-07-04
 ### Added
 - Heroku Import and DB Snapshot tasks